fix: Attempt to fix the detection of a running process on window for the CI runner
